rename services, fix translation
